improv

+++ litegapps/litegapps_menu.sh
@@ -3,9 +3,9 @@
 #Litegapps controller
 #▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
 #▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
-#29-12-2020 16-03-2021
-litegapps_menu_version=1.3
-litegapps_menu_code=3
+#29-12-2020 08-06-2021
+litegapps_menu_version=1.4
+litegapps_menu_code=4
 #▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
 #base func
 #▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
@@ -17,7 +17,7 @@ B='\e[01;34m'		# BLUE TEXT
 V='\e[01;35m'		# VIOLET TEXT
 Bl='\e[01;30m'		# BLACK TEXT
 C='\e[01;36m'		# CYAN TEXT
-W='\e[01;37m'		# WHITE TEXT
+WHITE='\e[01;37m'		# WHITE TEXT
 BGBL='\e[1;30;47m'	# Background W Text Bl
 N='\e[0m'			# How to use (example): echo "${G}example${N}"
 ####functions
@@ -55,12 +55,48 @@ x86_64) ARCH=x86_64 ;;
 esac
 
 base=/data/litegapps
+BASE=$base
+BIN=$BASE/bin
 #system
-system=/system
+SYSTEM_DIR=$SYSDIR
+API=`getprop ro.build.version.sdk`
+SDK=$API
+ABI=`getprop ro.product.cpu.abi | cut -c-3`
+ABI2=`getprop ro.product.cpu.abi2 | cut -c-3`
+ABILONG=`getprop ro.product.cpu.abi`
+
+ARCH=arm
+ARCH32=arm
+IS64BIT=false
+if [ "$ABI" = "x86" ]; then ARCH=x86; ARCH32=x86; fi;
+if [ "$ABI2" = "x86" ]; then ARCH=x86; ARCH32=x86; fi;
+if [ "$ABILONG" = "arm64-v8a" ]; then ARCH=arm64; ARCH32=arm; IS64BIT=true; fi;
+if [ "$ABILONG" = "x86_64" ]; then ARCH=x64; ARCH32=x86; IS64BIT=true; fi;
+
+if [ -f /data/adb/modules/litegapps/module.prop ]; then
+MODE_INSTALL=MAGISK
+MODE_DESK="Magisk Module"
+SYSTEM_INSTALL=/data/adb/modules/litegapps/system
+else
+MODE_DESK="LiteGapps Module"
+MODE_INSTALL=REGULER
+SYSTEM_INSTALL=/system
+fi
+
 
 #▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
 #main func
 #▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒
+DEV_MODE(){
+	chmod 755 $BIN/zip
+	DEFAULT_DIR=$(pwd)
+	DEV_BASE=/data/media/0/lol/package
+	cd $DEV_BASE
+	$BIN/zip -r1 package.zip * >/dev/null
+	cp -pf $DEV_BASE/package.zip $BASE/download/$name.zip
+	rm -rf $DEV_BASE/package.zip
+	cd $DEFAULT_DIR
+	}
 menu_end(){
 	print " "
 	print "${C}1.Back     ${Y}2.Reboot"
@@ -76,25 +112,38 @@ menu_end(){
 	}
 
 install_package(){
-	input="$1"
+	local input="$1"
 	rm -rf $base/tmp
 	mkdir -p $base/tmp
 	print "- Extracting package"
 	print " "
-	$bin2/busybox unzip -o "$input" -d  $base/tmp/ >/dev/null
-	if [ -f $base/tmp/litegapps-install.sh ]; then
-	chmod 755 $base/tmp/litegapps-install.sh
-	. $base/tmp/litegapps-install.sh
+	chmod 755 $BIN/unzip
+	$BIN/unzip -o "$input" -d  $base/tmp/ >/dev/null
+	if [ $? -eq 0 ]; then
+		print "- Unzip ${G}[OK]${G}"
 	else
-	print "${R} this package litegapps-install.sh not found ! $G"
+		print "- Unzip ${R}[ERROR]${WHITE} <$input>"
+		print " "
+		print "${R}Package is corrupt !!"
+		return 1
+	fi
+	if [ -f $base/tmp/litegapps-install.sh ]; then
+		chmod 755 $base/tmp/litegapps-install.sh
+		. $base/tmp/litegapps-install.sh
+	else
+		print "${R} this package litegapps-install.sh not found ! $G"
 	fi
 	
 	name_package_module=`getp package.module $base/tmp/litegapps-prop`
 	[ ! -d $base/modules ] && mkdir -p $base/modules
 	[ -d $base/modules/$name_package_module ] && rm -rf $base/modules/$name_package_module
 	[ ! -d $base/modules/$name_package_module ] && mkdir -p $base/modules/$name_package_module
-	for move_package in litegapps-prop litegapps-list litegapps-uninstall.sh litegapps-restore.sh module.prop; do
-		[ -f $base/tmp/$move_package ] && cp -pf $base/tmp/$move_package $base/modules/$name_package_module/
+	for move_package in litegapps-prop litegapps-list litegapps-uninstall.sh litegapps-restore.sh module.prop list-debloat backup; do
+		if [ -f $base/tmp/$move_package ]; then
+			cp -pf $base/tmp/$move_package $base/modules/$name_package_module/
+		elif [ -d $base/tmp/$move_package ]; then
+			cp -af $base/tmp/$move_package $base/modules/$name_package_module/
+		fi
 	done
 	rm -rf $base/tmp
 	}
@@ -112,7 +161,7 @@ if [ -f $base/modules/$name/litegapps-uninstall.sh ]; then
 	print "1.Install"
 	print "2.Uninstall"
 	print
-	echo -n "${G}Select Menu :${W} "
+	echo -n "${G}Select Menu :${WHITE} "
 	read menuin
 	case $menuin in
 	1)
@@ -133,40 +182,37 @@ modeselect=install
 fi
 
 if [ "$modeselect" = "install" ]; then
-clear
-printmid "${C}Install packages${G}"
-print
-[ -d $base2/download ] && rm -rf $base2/download
-test ! -d $base2/download && mkdir -p $base2/download
-print "- Download package"
-#cp -pf /sdcard/asw/package/package.zip $base2/download/$name.zip
-$bin2/curl -L -o $base2/download/$name.zip $url 2>/dev/null & spinner "- Downloading" 2>/dev/null
-ZIP_TEST="$(file -b $base2/download/$name.zip | head -1 | cut -d , -f 1)"
-case "$ZIP_TEST" in
-    "Zip archive data") 
-    ZIP_STATUS="file is not corrupt"
-    ;;
-    *)
-    print "${R} !!! File Is Corrupt !!!"
-    print " "
-    print "${G}* Tips"
-    print " "
-    print " Please check your internet connection and try again${W}"
-    print " "
-    sleep 6s
-    return 1
-    ;;
-esac
-
-install_package $base2/download/$name.zip
-rm -rf $base/download
+	clear
+	printmid "${C}Install packages${G}"
+	print
+	[ -d $base2/download ] && rm -rf $base2/download
+	test ! -d $base2/download && mkdir -p $base2/download
+	print "- Download package"
+	#DEV_MODE
+	$BIN/curl -L -o $base2/download/$name.zip $url
+	
+	if [ $? -eq 0 ]; then
+		print "${WHITE}- Downloading status : ${G}[OK]${WHITE}"
+		install_package $base2/download/$name.zip
+		rm -rf $base/download
+	else
+    	print "${R} !!! Downloading Package Failed !!!"
+    	print " "
+    	print "${G}* Tips"
+    	print " "
+    	print " Please check your internet connection and try again${WHITE}"
+    	print " "
+    fi
 elif [ "$modeselect" = "uninstall" ]; then
-clear
-printmid "${C}Uninstall Packages${G}"
-print
-[ -f $base/modules/$name/litegapps-uninstall.sh ] && chmod 755 $base/modules/$name/litegapps-uninstall.sh && . $base/modules/$name/litegapps-uninstall.sh
-[ -d $base/modules/$name ] && rm -rf $base/modules/$name
-print
+	clear
+	printmid "${C}Uninstall Packages${G}"
+	print
+	if [ -f $base/modules/$name/litegapps-uninstall.sh ]; then
+		chmod 755 $base/modules/$name/litegapps-uninstall.sh
+		. $base/modules/$name/litegapps-uninstall.sh
+	fi
+	[ -d $base/modules/$name ] && rm -rf $base/modules/$name
+	print
 fi
 
 menu_end
@@ -181,47 +227,35 @@ menu_download(){
 	clear
 	printmid "${C}Packages List${G}"
 	print
-	print "1.Wellbeing"
-	print "2.Vanced Manager"
-	print "3.Sound Picker"
-	print "4.Goole Play Games"
-	print "5.Carrier Service ${Y}(Recomended)${G}"
-	print "6.Google Location History ${Y}(Recomended)${G}"
-	print "7.Setup Wizard ${R}(BETA)${G}"
-	print "8.about"
-	print "9.Exit"
+	print "1.Google Services Framework"
+	print "2.GMS"
+	print "3.Play Store"
+	print "4.Setup Wizard"
+	print "5.About"
+	print "6.Exit"
 	print
 	echo -n " Select List Menu : ${V}"
 	read dmenu
 	case $dmenu in
 	1)
-	download_file Wellbeing https://sourceforge.net/projects/litegapps/files/database/$ARCH/$SDK/Wellbeing.zip/download
+	download_file GoogleServiceFramework https://sourceforge.net/projects/litegapps/files/addon/$ARCH/$SDK/GoogleServiceFramework.zip/download
 	;;
 	2)
-	download_file VancedManager https://sourceforge.net/projects/litegapps/files/database/$ARCH/$SDK/VancedManager.zip/download
+	download_file GmsCore https://sourceforge.net/projects/litegapps/files/addon/$ARCH/$SDK/GmsCore.zip/download
 	;;
 	3)
-	download_file SoundPicker https://sourceforge.net/projects/litegapps/files/database/$ARCH/$SDK/SoundPicker.zip/download
+	download_file PlayStore https://sourceforge.net/projects/litegapps/files/addon/$ARCH/$SDK/PlayStore.zip/download
 	;;
 	4)
-	download_file PlayGames https://sourceforge.net/projects/litegapps/files/database/$ARCH/$SDK/PlayGames.zip/download
+	download_file SetupWizard https://sourceforge.net/projects/litegapps/files/addon/$ARCH/$SDK/SetupWizard.zip/download
 	;;
 	5)
-	download_file CarrierServices https://sourceforge.net/projects/litegapps/files/database/$ARCH/$SDK/CarrierServices.zip/download
-	;;
-	6)
-	download_file GoogleLocationHistory https://sourceforge.net/projects/litegapps/files/database/$ARCH/$SDK/GoogleLocationHistory.zip/download
-	;;
-	7)
-	download_file SetupWizard https://sourceforge.net/projects/litegapps/files/database/$ARCH/$SDK/SetupWizard.zip/download
-	;;
-	8)
 	clear
 	printmid "${C}About${G}"
 	print
 	print "This is a tool to download manual Litegapps packages."
 	print
-	printmid "${Y}Problem Solving${W}"
+	printmid "${Y}Problem Solving${WHITE}"
 	print "1.make sure you have a good internet connection"
 	print "2.make sure you are using the latest version of litegapps controller"
 	print
@@ -229,8 +263,11 @@ menu_download(){
 	print
 	menu_end
 	;;
-	9)
+	6)
 	break
+	;;
+	*)
+	error "Please select number in menu list" 
 	;;
 	esac
 	done
@@ -285,11 +322,12 @@ print " "
 dirpackage=/data/media/0/Android/litegapps/package
 numzip=0
 for IZIP in $(ls -1 $dirpackage); do
-	ZIP_TEST="$(file -b $dirpackage/$IZIP | head -1 | cut -d , -f 1)"
-	if [ -f $dirpackage/$IZIP ] && [ "$ZIP_TEST" = "Zip archive data" ]; then
+	if [ -f $dirpackage/$IZIP ]; then
 		numzip2=$(((numzip+1)))
 		numzip=$numzip2
 		install_package $dirpackage/$IZIP
+	else
+		print "${R} This not zip file : <$dirpackage/$IZIP>"
 	fi
 done
 print
@@ -305,7 +343,7 @@ while true; do
 if [ -f $CONFIGDIR/backup_restore ]; then
 backup_restore="${G}1.Backup And Restore = ON${G}"
 else
-backup_restore="${G}1.Backup And Restore = ${W}OFF${G}"
+backup_restore="${G}1.Backup And Restore = ${WHITE}OFF${G}"
 fi
 clear
 printmid "${C}Settings${G}"
@@ -343,26 +381,28 @@ menu_about(){
 	print "Telegram channel : https://t.me/litegapps"
 	print " "
 	print
-	printmid "${Y}Download Package${W}"
+	printmid "${Y}Download Package${WHITE}"
 	print " "
 	print " "
 	printmid "${Y}Tweaks${G}"
 	print ""
 	print " "
-	printmid "${Y}Install ZIP Packages${W}"
+	printmid "${Y}Install ZIP Packages${WHITE}"
 	print " You can install even more than one zip package!  offline.  put package.zip into /sdcard/Android/litegapps/package"
 	print " "
 	printmid "${Y}Settings${G}"
 	print "Settings "
 	print " "
-	printmid "${Y}Updater${W}"
-	print " "
+	printmid "${Y}Updater${WHITE}"
+	print " litegapps menu update"
 	print " "
 	menu_end
 	}
 while true; do
 clear
 printmid "${C}Litegapps Menu${G}"
+print
+print " ${WHITE} Mode : ${Y}$MODE_DESK ${G}"
 print
 print "1.Download package"
 print "2.Tweaks"
